+++ src/main/java/org/team1540/robot2019/commands/cargo_intake/LowerThenCargoIntake.java
@@ -1,15 +1,14 @@
-package org.team1540.robot2019.commands.groups;
+package org.team1540.robot2019.commands.cargo_intake;
 
 import edu.wpi.first.wpilibj.command.CommandGroup;
 import org.team1540.robot2019.Robot;
-import org.team1540.robot2019.commands.cargo.AutoIntakeCargo;
 import org.team1540.robot2019.commands.elevator.MoveElevatorToZero;
 import org.team1540.robot2019.commands.wrist.WristDown;
 import org.team1540.rooster.util.SimpleCommand;
 
-public class IntakeSequence extends CommandGroup {
+public class LowerThenCargoIntake extends CommandGroup {
 
-    public IntakeSequence() {
+    public LowerThenCargoIntake() {
         addParallel(new SimpleCommand("alaska's", Robot.hatch::slideIn, Robot.hatch));
         addSequential(new MoveElevatorToZero());
         addParallel(new AutoIntakeCargo());
