+++ src/main/java/org/team1540/robot2019/commands/climber/ResetClimber.java
@@ -1,8 +1,7 @@
-package org.team1540.robot2019.commands.groups;
+package org.team1540.robot2019.commands.climber;
 
 import edu.wpi.first.wpilibj.command.CommandGroup;
 import org.team1540.robot2019.Tuning;
-import org.team1540.robot2019.commands.climber.MoveArmsToPosition;
 import org.team1540.robot2019.commands.elevator.MoveElevatorToZero;
 
 public class ResetClimber extends CommandGroup {
