+++ src/main/java/org/team1540/robot2019/commands/elevator/MoveElevatorToZero.java
@@ -9,6 +9,6 @@ public class MoveElevatorToZero extends CommandGroup {
   public MoveElevatorToZero() {
     addSequential(new MoveElevatorToPosition(0));
 //    addSequential(new ZeroElevator());
-    addSequential(new SimpleCommand("Stop Elevator", Robot.elevator::stop, Robot.elevator));
+    addSequential(new SimpleCommand("Stop Elevator", Robot.elevator::prettyPleaseStopBothTheLeftAndTheRightElevatorMotorsSoWeDontWastePowerCauseWeDontHavePowerManagement, Robot.elevator));
   }
 }
