+++ src/main/java/org/team1540/robot2019/commands/hatch/GetHatchFloor.java
@@ -1,4 +1,4 @@
-package org.team1540.robot2019.commands.groups;
+package org.team1540.robot2019.commands.hatch;
 
 import edu.wpi.first.wpilibj.command.Command;
 import edu.wpi.first.wpilibj.command.CommandGroup;
@@ -13,7 +13,7 @@ public class GetHatchFloor extends CommandGroup {
     addSequential(new SimpleCommand("in", Robot.hatch::slideIn, Robot.hatch));
     addSequential(new SimpleCommand("close", Robot.hatch::release, Robot.hatch));
     addSequential(new MoveElevatorToZero());
-      addSequential(new WristDown());
+    addSequential(new WristDown());
     addSequential(new SimpleCommand("out", Robot.hatch::slideOut, Robot.hatch));
     addSequential(new Command() {
       @Override
