+++ src/main/java/org/team1540/robot2019/commands/hatch/PlaceHatchThenDown.java
@@ -1,10 +1,9 @@
-package org.team1540.robot2019.commands.groups;
+package org.team1540.robot2019.commands.hatch;
 
 import edu.wpi.first.wpilibj.command.CommandGroup;
 import edu.wpi.first.wpilibj.command.WaitCommand;
 import org.team1540.robot2019.Tuning;
 import org.team1540.robot2019.commands.elevator.MoveElevatorToZero;
-import org.team1540.robot2019.commands.hatch.PlaceHatch;
 
 public class PlaceHatchThenDown extends CommandGroup {
