+++ src/main/java/org/team1540/robot2019/commands/self_tests/HatchSelfTest.java
@@ -1,7 +1,9 @@
-package org.team1540.robot2019.commands.hatch;
+package org.team1540.robot2019.commands.self_tests;
 
 import edu.wpi.first.wpilibj.command.CommandGroup;
 import org.apache.log4j.Logger;
+import org.team1540.robot2019.commands.hatch.PlaceHatch;
+import org.team1540.robot2019.commands.hatch.PrepGetHatch;
 import org.team1540.rooster.util.SimpleCommand;
 
 public class HatchSelfTest extends CommandGroup {
