+++ src/main/java/org/team1540/robot2019/commands/self_tests/SelfTest.java
@@ -1,11 +1,7 @@
-package org.team1540.robot2019.commands.groups;
+package org.team1540.robot2019.commands.self_tests;
 
 import edu.wpi.first.wpilibj.command.CommandGroup;
 import org.apache.log4j.Logger;
-import org.team1540.robot2019.commands.cargo.CargoIntakeSelfTest;
-import org.team1540.robot2019.commands.elevator.ElevatorSelfTest;
-import org.team1540.robot2019.commands.hatch.HatchSelfTest;
-import org.team1540.robot2019.commands.wrist.WristSelfTest;
 import org.team1540.rooster.util.SimpleCommand;
 
 public class SelfTest extends CommandGroup {
