+++ src/main/java/org/team1540/robot2019/commands/self_tests/WristSelfTest.java
@@ -1,8 +1,10 @@
-package org.team1540.robot2019.commands.wrist;
+package org.team1540.robot2019.commands.self_tests;
 
 import edu.wpi.first.wpilibj.command.CommandGroup;
 import edu.wpi.first.wpilibj.command.TimedCommand;
 import org.apache.log4j.Logger;
+import org.team1540.robot2019.commands.wrist.WristDown;
+import org.team1540.robot2019.commands.wrist.WristUpOrHold;
 import org.team1540.rooster.util.SimpleCommand;
 
 public class WristSelfTest extends CommandGroup {
