+++ src/main/java/org/team1540/robot2019/subsystems/Elevator.java
@@ -116,7 +116,7 @@ public class Elevator extends Subsystem {
         elevatorB.setIdleMode(mode);
     }
 
-    public void stop() {
+    public void prettyPleaseStopBothTheLeftAndTheRightElevatorMotorsSoWeDontWastePowerCauseWeDontHavePowerManagement() {
         setRaw(0);
     }
 }
