+++ src/main/java/org/team1540/robot2019/tuners/ClimberTuningRobot.java
@@ -8,8 +8,8 @@ import org.team1540.robot2019.Hardware;
 import org.team1540.robot2019.Robot;
 import org.team1540.robot2019.ShuffleboardDisplay;
 import org.team1540.robot2019.Tuning;
+import org.team1540.robot2019.commands.climber.PrepareForClimb;
 import org.team1540.robot2019.commands.climber.RaiseUpGyroAssist;
-import org.team1540.robot2019.commands.groups.PrepareForClimb;
 import org.team1540.robot2019.subsystems.Climber;
 import org.team1540.robot2019.subsystems.Drivetrain;
 import org.team1540.robot2019.subsystems.Elevator;
