+++ src/main/java/org/team1540/robot2019/tuners/ElevatorTuningRobot.java
@@ -34,7 +34,7 @@ public class ElevatorTuningRobot extends TimedRobot {
     }
 
     public void teleopInit() {
-        Robot.elevator.stop();
+        Robot.elevator.prettyPleaseStopBothTheLeftAndTheRightElevatorMotorsSoWeDontWastePowerCauseWeDontHavePowerManagement();
     }
 
     @Override
